不要なコメントの削除、記事カードを共通化
